questao_2_chatbot/requirements.txt
# Questão 2 – Chatbot com IA Generativa (Langchain + OpenAI)

# Frameworks de orquestração e LLM
langchain==0.3.0
langchain-openai==0.2.0

# SDK da OpenAI (dependência do langchain-openai)
openai>=1.59.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# Testes
pytest>=7.0.0
